get axis line in
