+++ src/styles/plot.scss
@@ -22,12 +22,6 @@ $rv-xy-plot-tooltip-padding: 7px 10px;
   }
 }
 
-.rv-xy-plot__axis__line {
-  fill: none;
-  stroke-width: 2px;
-  stroke: $rv-xy-plot-axis-line-color;
-}
-
 .rv-xy-plot__axis__tick__line {
   stroke: $rv-xy-plot-axis-line-color;
 }
